fixed some warnings from compiler

+++ smc/src/audio/audio.cpp
@@ -167,7 +167,7 @@ bool cAudio :: Init( void )
 	bool music = pPreferences->m_audio_music;
 
 	// if no change
-	if( numtimesopened && m_music_enabled == music && m_sound_enabled == sound && dev_frequency == pPreferences->m_audio_hz )
+	if( numtimesopened && m_music_enabled == music && m_sound_enabled == sound && (unsigned)dev_frequency == pPreferences->m_audio_hz )
 	{
 		return 1;
 	}
@@ -211,7 +211,7 @@ bool cAudio :: Init( void )
 		else
 		{
 			// different frequency
-			if( pPreferences->m_audio_hz != dev_frequency )
+			if( pPreferences->m_audio_hz != (unsigned)dev_frequency )
 			{
 				printf( "Warning : different frequency got %d but requested %d\n", dev_frequency, pPreferences->m_audio_hz );
 			}
